requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0

# Azure SDK
azure-identity==1.15.0
azure-mgmt-resource==23.0.1
azure-mgmt-monitor==6.0.2
azure-core==1.29.6

# AI/ML dependencies
openai==1.12.0
semantic-kernel==1.0.3

# Web framework for API
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1
